requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pillow>=10.0.0
python-multipart>=0.0.6
ollama>=0.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0